recsys/gcp/vertex_ai/containers/ranking/requirements.txt
flask>=2.0.0
google-cloud-aiplatform>=1.79.0,<2.0.0
google-cloud-storage>=2.0.0
google-cloud-bigquery>=3.0.0
pandas>=1.3.0
xgboost>=2.1.4
numpy>=1.21.0
scikit-learn>=1.0.0
python-dotenv>=1.0.0
vertexai>=0.3.0
uuid>=1.30
gunicorn>=21.0.0